Conflicts:
	docroot/static/js/main.js
	lib/RPG/C/Shop.pm
	lib/RPG/Schema.pm
	root/shop/standard_items.html

+++ lib/RPG/C/Building.pm
@@ -11,12 +11,42 @@ use List::Util qw(shuffle);
 
 sub auto : Private {
     my ($self, $c) = @_;
-    
+        
     if ($c->stash->{party}->in_combat) {
         croak "Can't manage buildings while in combat";   
     }
+        
+    $c->stash->{location} = $c->stash->{town} ? $c->stash->{town}->location : $c->stash->{party_location};
+    
+	$c->stash->{building} = $c->stash->{location}->building;
     
     return 1;
+}
+
+sub get_valid_groups {
+    my ($self, $c) = @_;
+    
+    my @groups;
+    
+	if ($c->stash->{town}) {
+	    my $mayor_cg = $c->stash->{town}->mayor->creature_group;
+	    push @groups, $mayor_cg if $mayor_cg;
+        
+        if ($c->stash->{party}->land_id == $c->stash->{town}->land_id) {
+            push @groups, $c->stash->{party};
+        }
+	}
+	else {
+	    @groups = ($c->stash->{party});
+	    
+	    my $garrison;
+	    if ($garrison = $c->stash->{party}->location->garrison and $garrison->party_id == $c->stash->{party}->id) {
+	        push @groups, $garrison;
+	    }
+	}
+	
+	return @groups;
+	
 }
 
 sub get_party_resources {
@@ -25,8 +55,15 @@ sub get_party_resources {
 	#  Get the list of resources owned by the current party.
 	my %resources;
 	
-	my @party_equipment = $c->stash->{party}->get_equipment(qw(Resource));
-	foreach my $resource (@party_equipment) {
+	my @equipment;
+	my @groups = $self->get_valid_groups($c, $c->stash->{town});
+	
+	foreach my $group (@groups) {
+	    next unless $group;
+	    push @equipment, $group->get_equipment('Resource');
+	}
+	
+	foreach my $resource (@equipment) {
         $resources{$resource->item_type->item_type} += $resource->variable('Quantity') // 0;	
 	}
 	
@@ -42,7 +79,7 @@ sub construct : Local {
         }
     );
     
-    my %party_resources = $self->get_party_resources($c);
+    my %party_resources = $self->get_party_resources($c, $c->stash->{town});
     
     my %resources = map { $_->item_type => $_ } $c->model('DBIC::Item_Type')->search(
         {
@@ -51,7 +88,12 @@ sub construct : Local {
         {
             join => 'category',
         }
-    ); 
+    );
+    
+    my @groups = $self->get_valid_groups($c, $c->stash->{town});
+    
+    my %resources_needed = $building_type->cost_to_build(\@groups);
+    my $enough_resources = $building_type->enough_resources(\@groups, %party_resources);
     	
 	$c->forward('RPG::V::TT',
         [{
@@ -60,40 +102,41 @@ sub construct : Local {
             	party => $c->stash->{party},
                 building_type => $building_type,
                 party_resources => \%party_resources,
-                enough_resources => $building_type->enough_resources(%party_resources),
+                enough_resources => $enough_resources,
                 resources => \%resources,
+                resources_needed => \%resources_needed,
+                town => $c->stash->{town},
             },
         }]
     );    
 }
 
 sub build : Local {
-    my ($self, $c) = @_; 
+    my ($self, $c) = @_;
     
-	my @current_building = $c->stash->{party_location}->building;
-    if (@current_building) {
+    if ($c->stash->{building}) {
         $c->stash->{error} = "There's already a building in this sector!";
         $c->detach('/panel/refresh');
     }   
+    
+    my $town = $c->stash->{town};
     
     my $building_type = $c->model('DBIC::Building_Type')->find(
         {
             level => 1,
         }
     );
+    
+    my @groups = $self->get_valid_groups($c, $town);
 
     if (! $building_type->enough_turns($c->stash->{party})) {
         $c->stash->{error} = "You don't have enough turns to construct the building";
         $c->detach('/panel/refresh');
     }
    
-	my %resources_needed = (
-       'Clay'  => $building_type->clay_needed,
-       'Iron'  => $building_type->iron_needed,
-       'Wood'  => $building_type->wood_needed,
-       'Stone' => $building_type->stone_needed,
-	);
-	if (! $c->stash->{party}->consume_items('Resource', %resources_needed) ) {
+	my %resources_needed = $building_type->cost_to_build(\@groups);
+	
+	if (! $building_type->consume_items(\@groups, %resources_needed) ) {
 		$c->stash->{error} = "Your party does not have the resources needed to create this building";
 		$c->detach('/panel/refresh');			
 	}    
@@ -101,10 +144,10 @@ sub build : Local {
 	#  Create the building.
 	my $building = $c->model('DBIC::Building')->create(
 		{
-			land_id => $c->stash->{party_location}->land_id,
+			land_id => $c->stash->{location}->land_id,
 			building_type_id => $building_type->id,
-			owner_id => $c->stash->{party}->id,
-			owner_type => "party",
+			owner_id => $town ? $town->id : $c->stash->{party}->id,
+			owner_type => $town ? 'town' : 'party',
 			name => $building_type->name,
 			
 			#  For now, partial construction not allowed, so we use all the materials up front
@@ -115,53 +158,49 @@ sub build : Local {
 			'labor_needed' => 0,
 		}
 	);   
-	
-	$c->forward('change_building_ownership', [$building]);
-
-	$c->model('DBIC::Party_Messages')->create(
-		{
-			message => "We created a " . $building_type->name . " at " . $c->stash->{party}->location->x . ", "
-			 . $c->stash->{party}->location->y,
-			alert_party => 0,
-			party_id => $c->stash->{party}->id,
-			day_id => $c->stash->{today}->id,
-		}
-	);
 
 	$c->stash->{party}->turns($c->stash->{party}->turns - $building_type->turns_needed($c->stash->{party}));
 	$c->stash->{party}->update;
 	
-	my $message = $c->forward( '/quest/check_action', [ 'constructed_building', $building ] );
+	if (! $town) {	
+    	$c->forward('change_building_ownership', [$building]);
+    
+    	$c->model('DBIC::Party_Messages')->create(
+    		{
+    			message => "We created a " . $building_type->name . " at " . $c->stash->{party}->location->x . ", "
+    			 . $c->stash->{party}->location->y,
+    			alert_party => 0,
+    			party_id => $c->stash->{party}->id,
+    			day_id => $c->stash->{today}->id,
+    		}
+    	);
+   	
+    	my $message = $c->forward( '/quest/check_action', [ 'constructed_building', $building ] );
+    	
+    	push @$message, "Building created";
+    	
+    	$c->stash->{panel_messages} = $message if @$message;
+    		
+    	$c->forward('/map/refresh_current_loc');
 	
-	push @$message, "Building created";
-	
-	$c->stash->{panel_messages} = $message if @$message;
-		
-	$c->forward('/map/refresh_current_loc');
-	
-	$c->forward( '/panel/refresh', [[screen => 'close'], 'party_status', 'messages'] );
+ 	   $c->forward( '/panel/refresh', [[screen => 'close'], 'party_status', 'messages'] );
+	}
 }
 
 sub manage : Local {
     my ($self, $c) = @_;
     
-	#  Get a list of the currently built (or under construction) buildings owned by the party.
-	my @existing_buildings = $c->model('DBIC::Building')->search(
-    	{
-    		'land_id' => $c->stash->{party_location}->id,
-        	'owner_id' => $c->stash->{party}->id,
-        	'owner_type' => 'party'
-        },
-        {
-            order_by => 'labor_needed'
-        },
-	);
+    my $building = $c->stash->{building};
+    
+	croak "No buildings to upgrade\n" unless $building;
 	
-	croak "No buildings to upgrade\n" unless @existing_buildings;
+	croak "Not allowed to manage building" unless $building->allowed_to_manage($c->stash->{party});
 	
-	my ($building) = @existing_buildings;
+    my $town = $c->stash->{town};
 	
 	my $building_type = $building->building_type;
+	
+	my @groups = $self->get_valid_groups($c, $town);
 	
 	my $upgradable_to_type = $c->model('DBIC::Building_Type')->find(
 	   {
@@ -169,7 +208,11 @@ sub manage : Local {
 	   }
 	);
 	
-    my %party_resources = $self->get_party_resources($c);
+	my @upgrade_types = $c->model('DBIC::Building_Upgrade_Type')->search();
+	
+	my %upgrades_by_type_id = map { $_->type_id => $_ } $building->upgrades;
+	
+    my %party_resources = $self->get_party_resources($c, $town);
     
     my %resources = map { $_->item_type => $_ } $c->model('DBIC::Item_Type')->search(
         {
@@ -179,6 +222,14 @@ sub manage : Local {
             join => 'category',
         }
     );	
+    
+    my %resources_needed;
+    my $enough_resources = 0;
+    
+    if ($upgradable_to_type) {
+        %resources_needed = $upgradable_to_type->cost_to_build(\@groups);
+        $enough_resources = $upgradable_to_type ? $upgradable_to_type->enough_resources(\@groups, %party_resources) : 1;
+    }
 
 	$c->forward('RPG::V::TT',
         [{
@@ -188,8 +239,13 @@ sub manage : Local {
                 building_type => $building_type,
                 upgradable_to_type => $upgradable_to_type,
                 party_resources => \%party_resources,
-                enough_resources => $upgradable_to_type->enough_resources(%party_resources),
+                enough_resources => $enough_resources,
                 resources => \%resources,
+                resources_needed => \%resources_needed,
+                town => $town,
+                upgrade_types => \@upgrade_types,
+                upgrades_by_type_id => \%upgrades_by_type_id,
+                building_url_prefix => $c->stash->{building_url_prefix},
             },
         }]
     );     
@@ -198,22 +254,14 @@ sub manage : Local {
 sub upgrade : Local {
     my ($self, $c) = @_;
     
-	#  Get a list of the currently built (or under construction) buildings owned by the party.
-	my @existing_buildings = $c->model('DBIC::Building')->search(
-    	{
-    		'land_id' => $c->stash->{party_location}->id,
-        	'owner_id' => $c->stash->{party}->id,
-        	'owner_type' => 'party'
-        },
-        {
-            order_by => 'labor_needed'
-        },
-	);
+    my $building = $c->stash->{building};
 	
-	croak "No buildings to upgrade\n" unless @existing_buildings;
+	croak "No buildings to upgrade\n" unless $building;
 	
-	my ($building) = @existing_buildings;
+	croak "Not allowed to manage building" unless $building->allowed_to_manage($c->stash->{party});
 	
+    my $town = $c->stash->{town};
+		
 	my $building_type = $building->building_type;
 	
 	my $upgradable_to_type = $c->model('DBIC::Building_Type')->find(
@@ -224,19 +272,16 @@ sub upgrade : Local {
 	
 	croak "Building can't be upgraded\n" unless $upgradable_to_type;	
 	
-    if (! $upgradable_to_type->enough_turns($c->stash->{party})) {
+    if (! $town && ! $upgradable_to_type->enough_turns($c->stash->{party})) {
         $c->stash->{error} = "You don't have enough turns to upgrade the building";
         $c->detach('/panel/refresh');
-    }	
+    }
+    
+    my @groups = $self->get_valid_groups($c, $town);
 	
-	my %resources_needed = (
-       'Clay'  => $upgradable_to_type->clay_needed,
-       'Iron'  => $upgradable_to_type->iron_needed,
-       'Wood'  => $upgradable_to_type->wood_needed,
-       'Stone' => $upgradable_to_type->stone_needed,
-	);
+	my %resources_needed = $upgradable_to_type->cost_to_build(\@groups);
 	
-	if (! $c->stash->{party}->consume_items('Resource', %resources_needed) ) {
+	if (! $building_type->consume_items(\@groups, %resources_needed) ) {
 		$c->stash->{error} = "Your party does not have the resources needed to upgrade this building";
 		$c->detach('/panel/refresh');			
 	}
@@ -244,17 +289,71 @@ sub upgrade : Local {
 	$building->building_type_id($upgradable_to_type->id);
 	$building->update;
 	
-	$c->forward('change_building_ownership', [$building]);
+	if (! $c->stash->{no_refresh}) {	
+    	$c->forward('change_building_ownership', [$building]);    
+    
+    	$c->stash->{party}->turns($c->stash->{party}->turns - $building_type->turns_needed($c->stash->{party}));
+    	$c->stash->{party}->update;
+    	
+    	$c->stash->{panel_messages} = ["Building upgraded"];
+    		
+    	$c->forward('/map/refresh_current_loc');
+    	
+    	$c->forward( '/panel/refresh', [[screen => 'close'], 'party_status', 'messages'] );
+	}	
+}
 
-
-	$c->stash->{party}->turns($c->stash->{party}->turns - $building_type->turns_needed($c->stash->{party}));
+sub build_upgrade : Local {    
+    my ($self, $c) = @_;
+    
+    my $building = $c->stash->{building};
+    
+    croak "Not allowed to manage building" unless $building->allowed_to_manage($c->stash->{party});
+    
+    my $town = $c->stash->{town};
+	
+	my $upgrade = $c->model('DBIC::Building_Upgrade')->find_or_create(
+	   {
+	       'building_id' => $building->id,
+	       'type_id' => $c->req->param('upgrade_type_id'),
+	   },
+	   { 
+	       for => 'update',
+	       prefetch => 'type', 
+	   },	   
+    );
+    
+    if ($building->building_type->max_upgrade_level < $upgrade->level + 1) {
+        croak "Cannot upgrade pasts max base building upgrade level";
+    }
+    
+	my %resources_needed = %{ $upgrade->type->cost_to_upgrade($upgrade->level + 1) };
+	
+	my $turns_needed = delete $resources_needed{Turns};
+	my $gold_needed = delete $resources_needed{Gold};
+	
+	my @groups = $self->get_valid_groups($c, $town);
+	
+	if (! $upgrade->type->consume_items(\@groups, %resources_needed) ) {
+		$c->stash->{error} = "Your party does not have the resources needed to build this upgrade";
+		$c->detach('/panel/refresh');			
+	}
+	
+	if ($turns_needed > $c->stash->{party}->turns || $gold_needed > $c->stash->{party}->gold) {
+        croak "Not enough turns or gold to build upgrade";
+	}
+	
+	$c->stash->{party}->decrease_gold($gold_needed);
+	$c->stash->{party}->turns($c->stash->{party}->turns - $turns_needed);
 	$c->stash->{party}->update;
 	
-	$c->stash->{panel_messages} = ["Building upgraded"];
-		
-	$c->forward('/map/refresh_current_loc');
+	$upgrade->level($upgrade->level+1);
+	$upgrade->update;
 	
-	$c->forward( '/panel/refresh', [[screen => 'close'], 'party_status', 'messages'] );	
+	if (! $c->stash->{no_refresh}) {	
+	   $c->forward( '/panel/refresh', [[screen => 'building/manage'], 'party_status'] );
+	}
+	
 }
 
 sub seize : Local {
@@ -265,7 +364,7 @@ sub seize : Local {
 		croak "You can't seize building - your party level is too low";
 	}
 	
-	if ( $c->stash->{party_location}->garrison ) {
+	if ( $c->stash->{party_location}->garrison && $c->stash->{party_location}->garrison->party_id != $c->stash->{party}->id ) {
 	   croak "Can't seize a building with a garrison";   
 	}	
 
@@ -274,9 +373,9 @@ sub seize : Local {
         $c->detach('/panel/refresh');
     }
 
-	#   Grab the building list, report on each on seized.
-	my @existing_buildings = $c->stash->{party_location}->building;	
-	my ($building) = @existing_buildings;
+	my $building = $c->stash->{building};
+	
+	croak "Cannot raze a town's building" if $building->owner_type eq 'town';
 	
 	croak "No building to seize\n" unless $building;
 	
@@ -346,7 +445,7 @@ sub seize : Local {
 	
 	$c->stash->{panel_messages} = ['Building Seized'];
 	
-	$c->forward( '/panel/refresh', ['messages', 'party_status'] );
+	$c->forward( '/panel/refresh', ['messages', 'party_status'] ) unless $c->stash->{no_refresh};
 }
 
 sub raze : Local {
@@ -357,11 +456,11 @@ sub raze : Local {
 		croak "You can't raze building - your party level is too low";
 	}
 	
-	#   Grab the building list, report on each on seized.
-	my @existing_buildings = $c->stash->{party_location}->building;	
-	my ($building) = @existing_buildings;
+	my $building = $c->stash->{building};
 	
 	croak "No building to raze\n" unless $building;
+	
+	croak "Cannot raze a town's building" if $building->owner_type eq 'town';
 	
 	my $turns_to_raze = $building->building_type->turns_to_raze($c->stash->{party});
 		
@@ -371,7 +470,7 @@ sub raze : Local {
 	    $c->detach('/panel/refresh');
 	}
 	
-	if ( $c->stash->{party_location}->garrison ) {
+	if ( $c->stash->{party_location}->garrison && $c->stash->{party_location}->garrison->party_id != $c->stash->{party}->id  ) {
 	   croak "Can't raze a building with a garrison";   
 	}
 
@@ -433,7 +532,7 @@ sub raze : Local {
 	
 	$c->forward('/map/refresh_current_loc');
 
-    $c->forward( '/panel/refresh', [[screen => 'close'], 'messages'] );
+    $c->forward( '/panel/refresh', [[screen => 'close'], 'messages'] ) unless $c->stash->{no_refresh};;
 }
 
 sub cede : Local {
@@ -441,9 +540,7 @@ sub cede : Local {
     
     croak "You don't have a Kingdom" unless $c->stash->{party}->kingdom_id;
     
-	# Grab the building list
-	my @existing_buildings = $c->stash->{party_location}->building;	
-	my ($building) = @existing_buildings;
+	my $building = $c->stash->{building};
 	
 	croak "No building to cede\n" unless $building;
